# internal/delivery/http/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/users": {
            "get": {
                "description": "Возвращает список пользователей с фильтрами и пагинацией",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Список пользователей",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Номер страницы",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Размер страницы",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Фильтр по статусу",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Фильтр по принятию оферты",
                        "name": "accepted_offer",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по промокоду",
                        "name": "promocode",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Возраст от",
                        "name": "age_from",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Возраст до",
                        "name": "age_to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Пол",
                        "name": "gender",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Наличие привязки карты",
                        "name": "map_binding",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Дата с (unix или YYYY-MM-DD)",
                        "name": "date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Дата по (unix или YYYY-MM-DD)",
                        "name": "date_to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.UsersListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/update": {
            "post": {
                "description": "Обновляет данные пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Обновить пользователя",
                "parameters": [
                    {
                        "description": "Данные пользователя",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.UpdateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.Status"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "description": "Возвращает информацию о пользователе по идентификатору",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Получить пользователя",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.UserResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "admin_history_pkg_proto_gen_go.User": {
            "type": "object",
            "properties": {
                "accepted_offer": {
                    "type": "boolean"
                },
                "age": {
                    "$ref": "#/definitions/wrapperspb.Int64Value"
                },
                "created_at": {
                    "$ref": "#/definitions/timestamppb.Timestamp"
                },
                "gender": {
                    "$ref": "#/definitions/wrapperspb.StringValue"
                },
                "id": {
                    "type": "integer"
                },
                "language": {
                    "$ref": "#/definitions/wrapperspb.StringValue"
                },
                "map_binding": {
                    "$ref": "#/definitions/wrapperspb.BoolValue"
                },
                "paid": {
                    "type": "integer"
                },
                "promocode": {
                    "$ref": "#/definitions/wrapperspb.StringValue"
                },
                "ref_boss_id": {
                    "$ref": "#/definitions/wrapperspb.Int64Value"
                },
                "status": {
                    "type": "boolean"
                },
                "total": {
                    "type": "integer"
                },
                "unpaid": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "server.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "server.Status": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "ok": {
                    "type": "boolean"
                }
            }
        },
...
        "server.UpdateUserRequest": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/admin_history_pkg_proto_gen_go.User"
                }
            }
        },
        "server.UserResponse": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/admin_history_pkg_proto_gen_go.User"
                }
            }
        },
        "server.UsersListResponse": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/admin_history_pkg_proto_gen_go.User"
                    }
                }
            }
        },
        "timestamppb.Timestamp": {
            "type": "object",
            "properties": {
                "nanos": {
                    "description": "Non-negative fractions of a second at nanosecond resolution. Negative\nsecond values with fractions must still have non-negative nanos values\nthat count forward in time. Must be from 0 to 999,999,999\ninclusive.",
                    "type": "integer"
                },
                "seconds": {
                    "description": "Represents seconds of UTC time since Unix epoch\n1970-01-01T00:00:00Z. Must be from 0001-01-01T00:00:00Z to\n9999-12-31T23:59:59Z inclusive.",
                    "type": "integer"
                }
            }
        },
        "wrapperspb.BoolValue": {
            "type": "object",
            "properties": {
                "value": {
                    "description": "The bool value.",
                    "type": "boolean"
                }
            }
        },
...
        "wrapperspb.Int64Value": {
            "type": "object",
            "properties": {
                "value": {
                    "description": "The int64 value.",
                    "type": "integer"
                }
            }
        },
        "wrapperspb.StringValue": {
            "type": "object",
            "properties": {
                "value": {
                    "description": "The string value.",
                    "type": "string"
                }
            }
        }
    }
}